# specs/003-add-candidate-management/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Candidate Management API
  version: 1.0.0
paths:
  /candidates:
    get:
      summary: List all candidates
      responses:
        '200':
          description: List of candidates
    post:
      summary: Create a candidate (form or document)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                skills:
                  type: array
                  items:
                    type: string
                document:
                  type: string
                  format: binary
      responses:
        '201':
          description: Candidate created
  /candidates/{id}:
    get:
      summary: Get candidate by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Candidate details
    put:
      summary: Update candidate by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                skills:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Candidate updated
    delete:
      summary: Delete candidate by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Candidate deleted
  /documents/{id}:
    get:
      summary: Get document by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document details